+++ lib/kitchen/directions/bake_notes/bake_autotitled_notes.rb
@@ -7,15 +7,17 @@ module Kitchen
         bake_subtitle: true,
         cases: false,
         bake_exercises: false
-      })
+      }, except_ids: [])
         options.reverse_merge!(
           bake_subtitle: true,
           cases: false,
           bake_exercises: false
         )
+        except_id_set = Set.new(except_ids)
 
         book.notes.each do |note|
           next unless (note.classes & classes).any?
+          next if except_id_set.include?(note.id)
 
           bake_note(
             note: note, options: options)